Revert part of commit c8ac11304ad5ec66e3dcf4b96222a73cf3ee4b83 (file renaming)
