+++ alphastats/gui/utils/gpt_helper.py
